Major code cleanup and performance boost of MPI override functions.
Inlined test_MPI to avoid a function call in the (most common)
failure-free case. Added library initialization check in test_MPI,
and in request addition and deletion functions. Removed comm_dup
when initializing __fenix_g_original_comm: we do not want to create
a duplicated communicator, we just want to be able to check its
identity with '=='.