Minor fixes for "RAID 0" mode
First, change "RAID 0" to "RAID 1" - I was using the wrong number throughout.

Second, fix an issue with choosing partners. Previously, recovery was done through groups of 2
wherein each ndoe had a single other node that it stored resiliency data on and that stored resiliency data on it.
Now, each node has two partners, on on which to store data, and one who's data it stores. This allows RAID 1 to be used
even when a group has an odd number of members.

TODO: I think the previous method is more resilient. If two nodes fail, the previous method has a lesser chance of a
non-recoverable fail pattern. If this seems true, we should update to use the previous "recovery groups" method when
an even number of members are in a group.